// src/components/Shared/Layout/ColorPicker/Buttons/button-interfaces.tsx
export interface CatButtonsProps {
  handleClickHair: () => void;
  handleClickEyes: () => void;
}

export interface DudeButtonsProps {
  handleClickHair: () => void;
  handleClickBeard: () => void;
  handleClickGlasses: () => void;
  handleClickShirt: () => void;
}

export interface HearingAidButtonsProps {
  handleClickOuter: () => void;
  handleClickInner: () => void;
}

export interface FunkyButtonsProps {
  handleClickHair: () => void;
  handleClickUnderCut: () => void;
  handleClickGlasses: () => void;
  handleClickShirt: () => void;
  handleClickJewellery: () => void;
}